requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
discord.py==2.4.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
frozenlist==1.5.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.0
motor==3.3.2
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
Nuitka==2.5.9
ordered-set==4.1.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
phantomjs==1.4.1
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pymongo==4.6.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.18.3
yt-dlp==2025.2.19
zstandard==0.23.0
